Moving helper into a util
